Move snapshot_given into shared snapshot-test conftest
The ``snapshot_given`` decorator is a general-purpose helper for any
snapshot test that wants to assert on the falsifying-example output of
a Hypothesis property test, so it belongs in
``tests/snapshots/conftest.py`` alongside ``SNAPSHOT_SETTINGS`` /
``EXPLAIN_SETTINGS`` rather than inline in ``test_data_object.py``.
